backend/requirements.txt
fastapi 
transformers
newspaper3k
lxml
pydantic
pytest
